requirements.txt
arabic-reshaper==2.0.14
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
configparser==3.5.0
cycler==0.10.0
future==0.16.0
hazm==0.5.2
idna==2.6
libwapiti==0.2.1
matplotlib==2.1.1
nltk==3.0.5
numpy==1.13.3
persian-wordcloud==1.3.1
Pillow==5.0.0
pyparsing==2.2.0
pyshorteners==0.6.1
python-bidi==0.4.0
python-dateutil==2.6.1
python-telegram-bot==9.0.0
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
wordcloud==1.3.1
